Changing to use the language version of primitive types for null saftey, updating controllers and service layers

+++ src/main/java/space/knightdev/dghandicap/dao/player/PlayerHandicap.java
@@ -7,6 +7,6 @@ import java.util.UUID;
 @Data
 public class PlayerHandicap {
     private UUID courseId;
-    private int layoutId;
-    private int handicap;
+    private Integer layoutId;
+    private Double handicap;
 }
